+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Vue 2.5+ is now the minimum required version
 * **Removed** prompt dialog ``inputPlaceholder``, ``inputName`` and ``inputMaxlength`` props
 * Add ``disabled`` prop to tabItem
+* #359 Add ``mobile-modal`` prop to dropdown to enable/disable modal on mobile
 
 ## 0.5.4